feat(playground-ui): extract MainSidebarTrigger into composable component
Extract the sidebar toggle button (with tooltip, icon, and Ctrl+B keyboard
shortcut) from MainSidebarRoot into a standalone MainSidebarTrigger component.

This allows external consumers to position the trigger anywhere in their
layout rather than having it hardcoded inside the sidebar root.

- New MainSidebarTrigger component exported as MainSidebar.Trigger
- Removed footerSlot prop from MainSidebarRoot (unused by any consumer)
- Updated playground AppSidebar to render the trigger explicitly
