+++ packages/playground-ui/src/ds/components/MainSidebar/main-sidebar-root.tsx
@@ -1,33 +1,13 @@
-import { KeyboardIcon, PanelRightIcon } from 'lucide-react';
-import { useEffect } from 'react';
 import { useMainSidebar } from './main-sidebar-context';
-import { MainSidebarNavSeparator } from './main-sidebar-nav-separator';
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/ds/components/Tooltip';
 import { cn } from '@/lib/utils';
 
 export type MainSidebarRootProps = {
   children: React.ReactNode;
   className?: string;
-  footerSlot?: React.ReactNode;
 };
-export function MainSidebarRoot({ children, className, footerSlot }: MainSidebarRootProps) {
+export function MainSidebarRoot({ children, className }: MainSidebarRootProps) {
   const { state, toggleSidebar } = useMainSidebar();
   const isCollapsed = state === 'collapsed';
-
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.ctrlKey && event.key === 'b') {
-        event.preventDefault();
-        toggleSidebar();
-      }
-    };
-
-    document.addEventListener('keydown', handleKeyDown);
-
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown);
-    };
-  }, [toggleSidebar]);
 
   return (
     <div
@@ -42,41 +22,6 @@ export function MainSidebarRoot({ children, className, footerSlot }: MainSidebar
       )}
     >
       {children}
-
-      <div className="bg-surface1 grid sticky bottom-0 pb-3">
-        <MainSidebarNavSeparator />
-        <div className="flex items-center justify-end gap-1">
-          {!isCollapsed && footerSlot}
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <button
-                onClick={toggleSidebar}
-                className={cn(
-                  'inline-flex w-auto items-center text-neutral3 h-8 px-3 rounded-md',
-                  'hover:bg-surface4 hover:text-neutral5',
-                  'transition-all duration-normal ease-out-custom',
-                  'focus:outline-hidden focus:ring-1 focus:ring-accent1 focus:shadow-focus-ring',
-                  '[&_svg]:w-4 [&_svg]:h-4 [&_svg]:text-neutral3 [&_svg]:transition-transform [&_svg]:duration-normal',
-                )}
-                aria-label="Toggle sidebar"
-              >
-                <PanelRightIcon
-                  className={cn({
-                    'rotate-180': isCollapsed,
-                  })}
-                />
-              </button>
-            </TooltipTrigger>
-
-            <TooltipContent>
-              Toggle Sidebar
-              <div className="flex items-center gap-1 [&>svg]:w-[1em] [&>svg]:h-[1em]">
-                <KeyboardIcon /> Ctrl+B
-              </div>
-            </TooltipContent>
-          </Tooltip>
-        </div>
-      </div>
 
       <button
         onClick={toggleSidebar}
