+++ packages/playground-ui/src/ds/components/MainSidebar/main-sidebar.tsx
@@ -6,11 +6,13 @@ import { MainSidebarNavList } from './main-sidebar-nav-list';
 import { MainSidebarNavSection } from './main-sidebar-nav-section';
 import { MainSidebarNavSeparator } from './main-sidebar-nav-separator';
 import { MainSidebarRoot } from './main-sidebar-root';
+import { MainSidebarTrigger } from './main-sidebar-trigger';
 
 export { MainSidebarProvider, type SidebarState } from './main-sidebar-context';
 export { useMainSidebar, useMaybeSidebar } from './main-sidebar-context';
 export { type NavLink } from './main-sidebar-nav-link';
 export { type NavSection } from './main-sidebar-nav-section';
+export { MainSidebarTrigger } from './main-sidebar-trigger';
 
 export const MainSidebar = Object.assign(MainSidebarRoot, {
   Bottom: MainSidebarBottom,
@@ -20,4 +22,5 @@ export const MainSidebar = Object.assign(MainSidebarRoot, {
   NavHeader: MainSidebarNavHeader,
   NavList: MainSidebarNavList,
   NavSeparator: MainSidebarNavSeparator,
+  Trigger: MainSidebarTrigger,
 });
